+++ test/integration/test_doubler_and_spy.gd
@@ -48,3 +48,9 @@ class TestBoth:
 		inst.add_user_signal('signal_with_params', ['a', 'b'])
 		assert_true(_spy.was_called(inst, 'add_user_signal'), 'added first signal')
 		assert_true(_spy.was_called(inst, 'add_user_signal', ['signal_with_params', ['a', 'b']]), 'second signal added')
+
+	func test_can_spy_on_native_doubles():
+		var inst  = _doubler.partial_double_gdnative(Node2D).new()
+		inst.set_position(Vector2(20, 20))
+		assert_true(_spy.was_called(inst, 'set_position'))
+		assert_true(_spy.was_called(inst, 'set_position', [Vector2(20, 20)]))