+++ test/integration/test_test_stubber_doubler.gd
@@ -82,19 +82,52 @@ class TestBasics:
 		assert_eq(a.get_a(), 10)
 		assert_eq(anotherA.get_a(), null)
 
-class TestTestsSmartDoubleMethod:
+class TestIgnoreMethodsWhenDoubling:
 	extends "res://test/gut_test.gd"
 
-	var _gut = null
+	var _test_gut = null
+	var _test = null
+
+	func before_each():
+		_test_gut = Gut.new()
+		_test = Test.new()
+		_test.gut = _test_gut
+
+	func test_when_calling_with_path_it_sends_path_to_doubler():
+		var m_doubler = double(_utils.Doubler).new()
+		_test_gut._doubler = m_doubler
+		_test.ignore_method_when_doubling('one', 'two')
+		assert_called(m_doubler, 'add_ignored_method', ['one', 'two'])
+
+	func test_when_calling_with_loaded_script_the_path_is_sent_to_doubler():
+		var m_doubler = double(_utils.Doubler).new()
+		_test_gut._doubler = m_doubler
+		_test.ignore_method_when_doubling(load(DOUBLE_ME_PATH), 'two')
+		assert_called(m_doubler, 'add_ignored_method', [DOUBLE_ME_PATH, 'two'])
+
+	func test_when_calling_with_scene_the_script_path_is_sent_to_doubler():
+		var m_doubler = double(_utils.Doubler).new()
+		_test_gut._doubler = m_doubler
+		_test.ignore_method_when_doubling(load(DOUBLE_ME_SCENE_PATH), 'two')
+		assert_called(m_doubler, 'add_ignored_method', ['res://test/resources/doubler_test_objects/double_me_scene.gd', 'two'])
+
+	func test_when_ignoring_scene_methods_they_are_not_doubled():
+		_test.ignore_method_when_doubling(load(DOUBLE_ME_SCENE_PATH), 'return_hello')
+		var m_inst = _test.double(DOUBLE_ME_SCENE_PATH).instance()
+		m_inst.return_hello()
+		# since it is ignored it should not have been caught by the stubber
+		_test.assert_not_called(m_inst, 'return_hello')
+
+class TestTestsSmartDoubleMethod:
+	extends "res://test/gut_test.gd"
 	var _test = null
 
 	func before_all():
-		_gut = Gut.new()
 		_test = Test.new()
 		_test.gut = gut
 
 	func after_each():
-		_gut.get_stubber().clear()
+		gut.get_stubber().clear()
 
 	func test_when_passed_a_script_it_doubles_script():
 		var inst = _test.double(DOUBLE_ME_PATH).new()
@@ -137,6 +170,10 @@ class TestTestsSmartDoubleMethod:
 		assert_eq(inst.__gut_metadata_.path, INNER_CLASSES_PATH, 'check path')
 		assert_eq(inst.__gut_metadata_.subpath, 'InnerA', 'check subpath')
 
+	func test_can_double_native_classes():
+		var inst = _test.double(Node2D).new()
+		assert_not_null(inst)
+
 class TestPartialDoubleMethod:
 	extends "res://test/gut_test.gd"
 
@@ -157,9 +194,11 @@ class TestPartialDoubleMethod:
 		inst.set_value(10)
 		assert_eq(inst.get_value(), 10)
 
+	# TODO this test is tempramental.  It has something to do with the loading
+	# of the doubles I think.  Should be fixed.
 	func test_partial_double_scene():
 		var inst = _test.partial_double(DOUBLE_ME_SCENE_PATH).instance()
-		assert_eq(inst.return_hello(), 'hello')
+		assert_eq(inst.return_hello(), 'hello', 'sometimes fails, should be fixed.')
 
 	func test_partial_double_inner():
 		var inst = _test.partial_double(INNER_CLASSES_PATH, 'InnerA').new()
@@ -179,8 +218,21 @@ class TestPartialDoubleMethod:
 		assert_eq(inst.get_a(), null)
 
 	func test_can_spy_on_partial_doubles():
+		var pass_count = _test.get_pass_count()
 		var inst = _test.partial_double(DOUBLE_ME_PATH).new()
 		inst.set_value(10)
 		_test.assert_called(inst, 'set_value')
 		_test.assert_called(inst, 'set_value', [10])
-		assert_eq(_test.get_pass_count(), 2)
+		assert_eq(_test.get_pass_count(), pass_count + 2)
+
+	func test_can_stub_partial_doubled_native_class():
+		var inst = _test.partial_double(Node2D).new()
+		_test.stub(inst, 'get_position').to_return(-1)
+		assert_eq(inst.get_position(), -1)
+
+	func test_can_spy_on_partial_doubled_native_class():
+		var pass_count = _test.get_pass_count()
+		var inst = _test.partial_double(Node2D).new()
+		inst.set_position(Vector2(100, 100))
+		_test.assert_called(inst, 'set_position', [Vector2(100, 100)])
+		assert_eq(_test.get_pass_count(), pass_count + 1, 'tests have passed')